@@ -16,13 +16,15 @@
     "@angular/animations": "^17.0.2",
     "@angular/common": "^17.0.2",
     "@angular/compiler": "^17.0.2",
-    "@angular/core": "^17.0.2",
+    "@angular/core": "^17.3.8",
     "@angular/forms": "^17.0.2",
     "@angular/platform-browser": "^17.0.2",
     "@angular/platform-browser-dynamic": "^17.0.2",
     "@angular/router": "^17.0.2",
     "@capacitor/app": "5.0.7",
-    "@capacitor/core": "^5.7.1",
+    "@capacitor/camera": "^6.0.0",
+    "@capacitor/core": "^5.7.5",
+    "@capacitor/filesystem": "^6.0.0",
     "@capacitor/haptics": "5.0.7",
     "@capacitor/keyboard": "5.0.8",
     "@capacitor/preferences": "^6.0.0",
@@ -45,9 +47,10 @@
     "@angular/cli": "^17.0.0",
     "@angular/compiler-cli": "^17.0.2",
     "@angular/language-service": "^17.0.2",
-    "@capacitor/cli": "^5.7.1",
+    "@capacitor/cli": "^5.7.5",
     "@ionic/angular-toolkit": "^9.0.0",
     "@types/jasmine": "~5.1.0",
+    "@types/react": "^18.3.2",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "eslint": "^7.26.0",
